+++ tests/external_botocore/_test_bedrock_chat_completion.py
@@ -10,7 +10,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionSummary"},
             {
                 "id": None,  # UUID that varies with each run
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "transaction_id": None,
                 "span_id": "span-id",
@@ -35,7 +35,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionMessage"},
             {
                 "id": None,  # UUID that varies with each run
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "request_id": "660d4de9-6804-460e-8556-4ab2a019d1e3",
                 "span_id": "span-id",
@@ -54,7 +54,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionMessage"},
             {
                 "id": None,  # UUID that varies with each run
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "request_id": "660d4de9-6804-460e-8556-4ab2a019d1e3",
                 "span_id": "span-id",
@@ -75,7 +75,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionSummary"},
             {
                 "id": None,  # UUID that varies with each run
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "transaction_id": None,
                 "span_id": "span-id",
@@ -98,7 +98,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionMessage"},
             {
                 "id": "1234-0",
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "request_id": "3bf1bb6b-b6f0-4901-85a1-2fa0e814440e",
                 "span_id": "span-id",
@@ -117,7 +117,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionMessage"},
             {
                 "id": "1234-1",
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "request_id": "3bf1bb6b-b6f0-4901-85a1-2fa0e814440e",
                 "span_id": "span-id",
@@ -138,7 +138,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionSummary"},
             {
                 "id": None,  # UUID that varies with each run
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "transaction_id": None,
                 "span_id": "span-id",
@@ -161,7 +161,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionMessage"},
             {
                 "id": None,  # UUID that varies with each run
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "request_id": "c5188fb5-dc58-4cbe-948d-af173c69ce0d",
                 "span_id": "span-id",
@@ -180,7 +180,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionMessage"},
             {
                 "id": None,  # UUID that varies with each run
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "request_id": "c5188fb5-dc58-4cbe-948d-af173c69ce0d",
                 "span_id": "span-id",
