+++ tests/external_botocore/test_boto3_sns.py
